uiautomator2/requirements.txt
pytest
pytest-rerunfailures
allure-pytest
openpyxl